Ef to ui mapping

+++ Helper/ExtensionMethods.cs
@@ -252,7 +252,7 @@ namespace Vulnerator.Helper
         public static string ToFortifyThreshold(this string _string)
         {
             try
-            { return int.Parse(_string) >= 2.5 ? "High" : "Low"; }
+            { return double.Parse(_string) >= 2.5 ? "High" : "Low"; }
             catch (Exception exception)
             {
                 LogWriter.LogError($"Unable to convert raw status '{_string}' to a Fortify threshold.");
